Day 13 cleanup
Let's be honest, that one was rather clean already.

My biggest frustration here is the use of threading to transfer joystick
events to the input stream.  Possibly a byproduct of using streams at all.

In the end, this one should end up in the free monad IntCode interface,
but in the meantime I'm yearning for a select-like function that could
observe whether a new item has been produced in a lazy list.

Maybe I should go and ask StackOverflow someday.