IntCode: eradicate MonadFail
Its only use was to guard again this type of pattern match from failing:

  [Left op1,Left op2,Right out] <- readParams modes [In,In,Out]

Now parameter reading is done in a typesafe way where the redundancy
is eliminated:

  PReadVal opLeft :< PReadVal opRight :< PStore store :< () <- readParams modes

There remains to have a long enough modes list .  This shouldn't fail
(famous last words), but isn't statically guarded yet: all there is
is a call to error if readParams exhausts it.  Don't add an operation
with more than three parameters for now; I'll see if there's any point
in making that one a static list as well.

Now MonadFail is free to use for the actual relevant use case:
erroneous IntCode programs:
  * unknown opcode
  * writing to an immediate
  * reading/writing before address 0