Day 13: use IntCode's free monad interface
It works.

It was a lot of work to make it work.

It's not so very satisfying.
While I'm not depending anymore on the fact that the provided IntCode
requests input once per ball output (modulo start screen), it's
noticeably longer, and quite slower to boot.

This could probably go with a lot of INLINEs in IntCode, but it's
going to be something to keep an eye on.